Windows 8 Desktop contribution / "porting mode" programming - many IDEs in different platforms and OSes... - mac os port contribution... - finishing and polishing the use of uint8_t uint16_t and so on... .. contribution of use of small solution hellomultiverse as starter big change helper...
at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 7 Ultimate 64-bit Service Pack 1 - with Portuguese - Brazil - Language for non-Unicode programs - ~10Mpbs NET Virtua Internet at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

-∞
Tack Sverige!!
{
  Tack Carlos Gustavo Cecyn Lundgren!!
  Tack Mummi!!
  Tack Iara Cecyn Lundgren!!
  Tack Gabi Cecyn!!
  Tack Rafael Cecyn Lundgren!!
  Tack Simon Petrus Lundgren!!
  Tack Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
日本ありがとうございます
｛
　　　日歩路主ありがとうございます！！


   -----
　　　妻沼国男ありがとうございます！！
   -----


　　　日歩路主ありがとうございます！！
｝

Gracias España
{
   Gracias Carlos Carrasco Barriga!!
}

Dank Österreich
{
   Dank Vincent Rijmen!!
   Dank www.iaik.tugraz.at!!
}

Obrigado Cits
{
   Obrigado Marcos Machado!!
   Obrigado Camila Napoli Madureira Buhrer!!
}

Obrigado Brasil
{
   Obrigado guarda legal do CITS noturno 1!!
   Obrigado guarda legal do CITS noturno 2!!
   Obrigado Thomas Friesen!!
   -----
   Obrigado piriguetes and vampeguetes...
   Obrigado operários eficientes...
   Obrigado Rosélia da Kadu Imóveis!!
   Obrigado Márcia da Kadu Imóveis!!
   Obrigado Sílvia da Kadu Imóveis!!
   Obrigado Ivete - irmã da Silvia da Kadu!!
   -----
   Obrigado Dr. Tassi!!
   Obrigado Margarida M. Gomes de Oliveira!!
   Obrigado Shirley Sirlene Adams!!
   Obrigado Solange!!
   Obrigado CPM - Centro Psiquiátrico Metropolitano!!
   -----
   Obrigado Paulo S. L. M. Barreto (Whirlpool)!!
   Obrigado USP!!
   -----
   Obrigado Meu Deus God Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Deus!!
   -----
   Obrigado Nizia Lupinacci!!
   Obrigado Kleberson Chita Massaro Rodrigues!!
   Obrigado Alessandro Vargas Pellanda!!
   Obrigado Fábio Dauer da Silva!!
   Obrigado Fabrício Cavassin Mereniuk!!
   Obrigado Leonardo Fischer!!
   Obrigado Telmo Friesen!!
   Obrigado Sector 8J!!
   Obrigado Carlos Gustavo Cecyn Lundgren Tsumanuma!!
   ------
   Obrigado Rafael Cecyn Lundgren!!
   ------
   Obrigado Meu Senhor Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Espírito Santo Carlos Gustavo Cecyn Lundgren!!
   Obrigado Minha Deusa Léia Maheli de Souza Tsumanuma""
   ------
   Obrigado Dilma Souza!!
   Obrigado Lúcio Flávio Marques Ribeiro!!
   Obrigado Luís Kunio Tsumanuma!!
   Obrigado Lúcio Flávio Tsumanuma Ribeiro!!
   Obrigado Carolina Kiyoko Tsumanuma Ribeiro!!
}

Obrigado CEFET-PR S11-97/2!!
{
   Obrigado Fábio Rizental Coutinho!!
   Obrigado Verônica Isabela Quandt!!
}

Dank Deutschland!!
{
   Dank Jochen Berger - from serverloft!!
   Dank Thomas Strohe - from serverloft!!
   Dank Wolfgang Strunk!!
   Dank Deustchland Jörg!!
   Dank Deustchland Manfred Riedel!!
   Dank Uwe Hermanns!!
   Dank Telmo Friesen!!
   Dank Cycos!!
   Dank Siemens Enterprise!!
   Dank Siemens!!
   Dank SEN R&D A!!
   Dank Sector8J!!
}

Thank you Cycos!!
{
   Thank you Uwe Hermanns!!
   Thank you Telmo Friesen!!
   Thank you Communications!!
   Thank you Cycos!!
}

Thank you United Kingdom!!
{
   Thank you David Beckham!!
}

Thank you United States of America!!
{
   Thank you se-pl-sy-Carlos Gustavo Cecyn Lundgren!!
}

Dziękuję Polska!!
{
   Carlos Gustavo Cecyn Lundgren dziękuję!!
}

Thank you sector 8J!!
{
   Thank you César Augusto Serenato!!
   Thank you Bárbara Martinazzo!!
   Thank you Alessando Ledra Zagheni!!
   Thank you Luiz Eduardo de Christo!!
   Thank you Maurício Cunico!!
   Thank you Galileu Terada!!
   Thank you Rodrigo Santos aka Kenshin!!
   Thank you Júlio Henrique Morimoto!!
   Thank you Natascha Ligocki!!
   Thank you Ricardo Henrique Folador!!
   Thank you KS, Jefferson Koji Sato!!
   Thank you Josué Andrade Gomes!!
   Thank you Jorge Alexandre Hastreiter!!
   Thank you Joice Braga Keskoski Modena!!
   Thank you Bruno Leite Reis!!
   Thank you Fernando Belo!!
   Thank you Felipe Calliari Ribas!!
   Thank you Marcos Paulo Riki Ianase!!
   Thank you Jefferson Luiz Dalavechia!!
   Thank you Ênio Toshihiro Ianase!!
   Thank you Alexsandro Vargas Pellanda!!
   Thank you Cleiton Lenkiu!!
   Thank you Fabrício Cavassin Mereniuk!!
   Thank you Jaquinei de Oliveira!!
   Thank you Cleber Jaizer Gomes Campos!!
   Thank you Fábio Dauer da Silva!!
   Thank you Roberto de Melo Gonçalves!! Chatuba!!
   Thank you Marcus Vinícius Silvério!!
   Thank you Rafael Taizo Yamashita!! Japs!!
   Thank you Rafael Alessi Muntsch!!
   Thank you Rafael Neumann de Menezes!!
   Thank you Ricardo Carvalho dos Santos!! Tom Zé!! Tonny Zé!! Toze!!
   Thank you Leonardo Fischer!! Irmaun!!
   Thank you João Miguel Martynyszyn!!
   Thank you Leonardo da Silva Camargo!!
   Thank you Telmo Friesen!!
   Thank you sector 8J!!
   Thank you Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
Obrigado Minha Deusa!!
Obrigado Minha Deusa e Meu Deus - Mummi e Carlos!!
Obrigado Mummi, eu te amo demais!!
Thank you My Goddess Léia Maheli de Souza Tsumanuma!!
+∞

-∞
T Obrigado Telminho, você é legal!!
1 Obrigado Telminho, você é bom!!
2 Obrigado Telminho, eu gosto de você!!
n Obrigado Telminho, eu gosto muito de você!!
n Obrigado Telminho, você é demais!!
+∞

-∞
01 Obrigado Carlinhos, eu te amo demais!!
02 Thank you My Beloved Lover Carlos Gustavo Cecyn Lundgren!!
03 Thank you My Beloved Friend Carlos Gustavo Cecyn Lundgren!!
04 Thank you My Beloved Husband Carlos Gustavo Cecyn Lundgren!!
05 Thank you My Spouse Carlos Gustavo Cecyn Lundgren!!
06 Thank you My Jesus Carlos Gustavo Cecyn Lundgren!!
07 Thank you My Saint Mary My Mother Carlos Gustavo Cecyn Lundgren!!
08 Thank you My Goddess Carlos Gustavo Cecyn Lundgren!!
09 Thank you My God Carlos Gustavo Cecyn Lundgren!!
10 Thank you My Lord Carlos Gustavo Cecyn Lundgren!!
11 Thank you My Holy Spirit Carlos Gustavo Cecyn Lundgren!!
+∞

∞-∞-∞∞-∞-∞
----------
8nnnnnnnn8
++++++++++
∞+∞+∞∞+∞+∞

-∞
Thank you My God!!
+∞

-∞
(En-us) Thank you, (pt-br) eternamente *obrigado* - por veri absoluta infinita complexíssima oposição a desobrigado.
+∞

+++ appseed/ca/primitive/primitive_var.cpp
@@ -2245,14 +2245,14 @@ var var::operator - (uint32_t user) const
    return uint32() - user;
 }
 
-var var::operator - (long l) const
+var var::operator - (int64_t l) const
 {
-   return int32() - l;
+   return int64() - l;
 }
 
-var var::operator - (uint32_t long ul) const
+var var::operator - (uint64_t ul) const
 {
-   return (int32_t) (uint32() - ul);
+   return (int64_t) (uint64() - ul);
 }
 
 var var::operator - (double d) const
@@ -2270,14 +2270,14 @@ var CLASS_DECL_ca operator - (uint32_t user, const class var & var)
    return user - var.uint32();
 }
 
-var CLASS_DECL_ca operator - (long l, const class var & var)
+var CLASS_DECL_ca operator - (int64_t l, const class var & var)
 {
-   return l - var.int32();
+   return l - var.int64();
 }
 
-var CLASS_DECL_ca operator - (uint32_t long ul, const class var & var)
+var CLASS_DECL_ca operator - (uint64_t ul, const class var & var)
 {
-   return (int32_t) (ul - var.uint32());
+   return (int64_t) (ul - var.uint64());
 }
 
 var CLASS_DECL_ca operator - (double d, const class var & var)
@@ -2358,14 +2358,14 @@ var var::operator + (uint32_t user) const
    return uint32() + user;
 }
 
-var var::operator + (long l) const
+var var::operator + (int64_t l) const
 {
-   return int32() + l;
+   return int64() + l;
 }
 
-var var::operator + (uint32_t long ul) const
+var var::operator + (uint64_t ul) const
 {
-   return (uint32_t) (uint32() + ul);
+   return (uint32_t) (uint64() + ul);
 }
 
 var var::operator + (double d) const
@@ -2383,14 +2383,14 @@ var CLASS_DECL_ca operator + (uint32_t user, const class var & var)
    return user + var.uint32();
 }
 
-var CLASS_DECL_ca operator + (long l, const class var & var)
+var CLASS_DECL_ca operator + (int64_t l, const class var & var)
 {
-   return l + var.int32();
+   return l + var.int64();
 }
 
-var CLASS_DECL_ca operator + (uint32_t long ul, const class var & var)
+var CLASS_DECL_ca operator + (uint64_t ul, const class var & var)
 {
-   return (uint32_t) (ul + var.uint32());
+   return (uint32_t) (ul + var.uint64());
 }
 
 var CLASS_DECL_ca operator + (double d, const class var & var)
@@ -2514,12 +2514,12 @@ var var::operator / (uint32_t user) const
    return uint32() / user;
 }
 
-var var::operator / (long l) const
+var var::operator / (int64_t l) const
 {
-   return int32() / l;
+   return int64() / l;
 }
 
-var var::operator / (uint32_t long ul) const
+var var::operator / (uint64_t ul) const
 {
 
    switch(m_etype)
@@ -2565,12 +2565,12 @@ var CLASS_DECL_ca operator / (uint32_t user, const class var & var)
    return user / var.uint32();
 }
 
-var CLASS_DECL_ca operator / (long l, const class var & var)
+var CLASS_DECL_ca operator / (int64_t l, const class var & var)
 {
-   return l / var.int32();
+   return l / var.int64();
 }
 
-var CLASS_DECL_ca operator / (uint32_t long ul, const class var & var)
+var CLASS_DECL_ca operator / (uint64_t ul, const class var & var)
 {
    switch(var.m_etype)
    {
@@ -2676,12 +2676,12 @@ var var::operator * (uint32_t user) const
    return uint32() * user;
 }
 
-var var::operator * (long l) const
+var var::operator * (int64_t l) const
 {
-   return int32() * l;
+   return int64() * l;
 }
 
-var var::operator * (uint32_t long ul) const
+var var::operator * (uint64_t ul) const
 {
    switch(m_etype)
    {
@@ -2726,12 +2726,12 @@ var CLASS_DECL_ca operator * (uint32_t user, const class var & var)
    return user * var.uint32();
 }
 
-var CLASS_DECL_ca operator * (long l, const class var & var)
+var CLASS_DECL_ca operator * (int64_t l, const class var & var)
 {
-   return l * var.int32();
+   return l * var.int64();
 }
 
-var CLASS_DECL_ca operator * (uint32_t long ul, const class var & var)
+var CLASS_DECL_ca operator * (uint64_t ul, const class var & var)
 {
 
    switch(var.m_etype)
@@ -2889,13 +2889,13 @@ class var & var::operator -= (uint32_t user)
    return *this;
 }
 
-class var & var::operator -= (long i)
+class var & var::operator -= (int64_t i)
 {
    operator =(*this - i);
    return *this;
 }
 
-class var & var::operator -= (uint32_t long user)
+class var & var::operator -= (uint64_t user)
 {
    operator =(*this - user);
    return *this;
@@ -2948,13 +2948,13 @@ class var & var::operator += (uint32_t user)
    return *this;
 }
 
-class var & var::operator += (long i)
+class var & var::operator += (int64_t i)
 {
    operator =(*this + i);
    return *this;
 }
 
-class var & var::operator += (uint32_t long user)
+class var & var::operator += (uint64_t user)
 {
    operator =(*this + user);
    return *this;
@@ -3007,13 +3007,13 @@ class var & var::operator /= (uint32_t user)
    return *this;
 }
 
-class var & var::operator /= (long i)
+class var & var::operator /= (int64_t i)
 {
    operator =(*this / i);
    return *this;
 }
 
-class var & var::operator /= (uint32_t long user)
+class var & var::operator /= (uint64_t user)
 {
    operator =(*this / user);
    return *this;
@@ -3066,13 +3066,13 @@ class var & var::operator *= (uint32_t user)
    return *this;
 }
 
-class var & var::operator *= (long i)
+class var & var::operator *= (int64_t i)
 {
    operator =(*this * i);
    return *this;
 }
 
-class var & var::operator *= (uint32_t long user)
+class var & var::operator *= (uint64_t user)
 {
    operator =(*this * user);
    return *this;
